+++ NEWS.md
@@ -1,11 +1,19 @@
+# semmcci 1.1.0
+
+## Minor
+
+* Added option to replace variances less than or equal to zero in `MC()`.
+* Use `lavaan::standardizedSolution()` in `MCStd()`.
+* Added option for parallel processing in `MC()` and `MCStd()`.
+ 
 # semmcci 1.0.1
 
-## Minor changes
+## Patch
 
 * Added a reference in the DESCRIPTION file.
 
 # semmcci 1.0.0
 
-## Major changes
+## Major
 
 * Initial CRAN submission.